Added Sample selection
The application now automatically choses samples to pose as positive
examples of the mapping. Only the samples are mapped to semantic types
and persisted on the database.

+++ src/main/java/br/uniriotec/ppgi/mapping/controller/sampler/SampleSizeCalculator.java
@@ -61,7 +61,7 @@ public class SampleSizeCalculator {
 	 * @return the sample size needed for a specific population
 	 */
 	public static double correctToFinitePopulation(double rawSampleSize, int population){
-		return (int)(rawSampleSize / (1 + ((rawSampleSize - 1)/population)));
+		return (rawSampleSize / (1 + ((rawSampleSize - 1)/population)));
 	}
 	
 }
